webapp/requirements.txt
Flask
torch
torchvision
Pillow
numpy
matplotlib
opencv-python-headless
albumentations
gunicorn
